Swap algo for smart ptrs

+++ cpp_shared_ptrs/input_receiver.cpp
@@ -13,8 +13,8 @@ const char DELIMITER = ',';
 int read_matrix_size_from_file(std::ifstream &inputFile);
 void validate_matrix(std::ifstream &inputFile, const int &size);
 std::ifstream create_file_stream(std::string &file_name);
-std::shared_ptr<long double*[]> allocate_matrix(const int& size);
-void populate_matrix(std::ifstream &inputFile, std::shared_ptr<long double*[]> matrix, const int& size);
+std::shared_ptr<double*[]> allocate_matrix(const int& size);
+void populate_matrix(std::ifstream &inputFile, std::shared_ptr<double*[]> matrix, const int& size);
 
 matrix::Matrix input::InputReceiverImpl::read_matrix_from_file(std::string file_name) {
     std::ifstream inputFile = create_file_stream(file_name);
@@ -23,7 +23,7 @@ matrix::Matrix input::InputReceiverImpl::read_matrix_from_file(std::string file_
     inputFile.close();
     inputFile.open(file_name);
 
-    std::shared_ptr<long double*[]> matrix(allocate_matrix(size));
+    std::shared_ptr<double*[]> matrix(allocate_matrix(size));
     populate_matrix(inputFile, matrix, size);
 
     inputFile.close();
@@ -32,17 +32,17 @@ matrix::Matrix input::InputReceiverImpl::read_matrix_from_file(std::string file_
     return result;
 }
 
-std::shared_ptr<long double*[]> allocate_matrix(const int& size) {
-    std::shared_ptr<long double*[]> result(new long double*[size]);
+std::shared_ptr<double*[]> allocate_matrix(const int& size) {
+    std::shared_ptr<double*[]> result(new double*[size]);
 
     for (int i = 0; i < size; i++) {
-        result[i] = new long double[size];
+        result[i] = new double[size];
     }
 
     return result;
 }
 
-void populate_matrix(std::ifstream &inputFile, std::shared_ptr<long double*[]> matrix, const int& size) {
+void populate_matrix(std::ifstream &inputFile, std::shared_ptr<double*[]> matrix, const int& size) {
     std::string line;
 
     for (int row = 0; row < size; row++) {
